requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.31
Werkzeug==3.0.3
reportlab==4.2.2
qrcode==7.4.2
Pillow==10.4.0
APScheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
